primitives gone, controllers using repositories, orika for bean converter

+++ src/main/java/ar/edu/uade/integracion/shop/entity/ItemDto.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public class ItemDto {
 
-    private int id;
+    private Integer id;
     private UserDto seller;
-    private float price;
+    private Double price;
     private String name;
     private String description;
     private List<String> photos;
@@ -15,11 +15,11 @@ public class ItemDto {
     private Category category;
 
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -31,11 +31,11 @@ public class ItemDto {
         this.seller = seller;
     }
 
-    public float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
